refactor: abstract bill summary with receipt detail helpers
